refactor fluidspaces.py from symlink to full file
this used to be a symlink to a file with actual code in it so that I
could make a copy of the file when it was in a good configuration before
I kept working, and if I needed to stop before it worked again I could
just point the symlink back to the last good file.

but I messed that up immediately because of laziness ("this should just
be a tiny change") so now i'm just naming a copy of the code the
symlink pointed to last using the name of the symlink.

this will be broken up into separate files soon anyway, i hope.