refactor versioning: use git tags not VERSION file
turned out to be very difficult to read a VERSION file from the root of
the project using both setup.py and modules in the main package.

talked to altendky on the #python IRC and decided to try setuptools_scm
to get the version number from git tags instead (he said it was working
well for him).  we'll see how it goes.